feat#49: renamed projects->project and Project->Manager. Now interface is referred to as projects.Manager.
