+++ internal/project/nav.go
@@ -1,4 +1,4 @@
-package projects
+package project
 
 import (
 	"encoding/json"
