+++ internal/project/nav_test.go
@@ -1,4 +1,4 @@
-package projects
+package project
 
 import (
 	"fmt"
